+++ Sofa3DevOpsTest/UC10.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Sofa3DevOpsTest
+{
+    public class UC10
+    {
+        public UC10() { }
+
+        [Fact]
+        public void TestItemFromReadyForTestingToTesting()
+        {
+
+        }
+        
+        [Fact]
+        public void TestItemFromReadyForTestingToTestingByNonTester()
+        {
+
+        }
+    }
+}
